+++ src/ext/java/org/opentripplanner/ext/transmodelapi/model/timetable/InterchangeType.java
@@ -44,9 +44,9 @@ public class InterchangeType {
           .name("priority")
           .description(
             "The transfer priority is used to decide where a transfer should " +
-            "happen, at the highest prioritized location. If the guarantied " +
-            "flag is set it take precedence priority. A guarantied ALLOWED " +
-            "transfer is preferred over a PREFERRED none-guarantied transfer."
+            "happen, at the highest prioritized location. If the guaranteed " +
+            "flag is set it take precedence priority. A guaranteed ALLOWED " +
+            "transfer is preferred over a PREFERRED none-guaranteed transfer."
           )
           .type(EnumTypes.INTERCHANGE_PRIORITY)
           .dataFetcher(env -> constraint(env).getPriority())
@@ -58,7 +58,7 @@ public class InterchangeType {
           .name("maximumWaitTime")
           .description(
             "Maximum time after scheduled departure time the connecting " +
-            "transport is guarantied to wait for the delayed trip. [NOT " +
+            "transport is guaranteed to wait for the delayed trip. [NOT " +
             "RESPECTED DURING ROUTING, JUST PASSED THROUGH]"
           )
           .type(Scalars.GraphQLInt)