+++ src/ext/java/org/opentripplanner/ext/vehiclerentalservicedirectory/VehicleRentalParameters.java
@@ -1,5 +1,6 @@
 package org.opentripplanner.ext.vehiclerentalservicedirectory;
 
+import java.time.Duration;
 import org.opentripplanner.updater.vehicle_rental.VehicleRentalUpdaterParameters;
 import org.opentripplanner.updater.vehicle_rental.datasources.params.VehicleRentalDataSourceParameters;
 
@@ -7,9 +8,9 @@ public class VehicleRentalParameters extends VehicleRentalUpdaterParameters {
 
   public VehicleRentalParameters(
     String configRef,
-    int frequencySec,
+    Duration frequency,
     VehicleRentalDataSourceParameters sourceParameters
   ) {
-    super(configRef, frequencySec, sourceParameters);
+    super(configRef, frequency, sourceParameters);
   }
 }