+++ src/main/java/org/opentripplanner/api/common/OTPExceptionMapper.java
@@ -41,9 +41,7 @@ public class OTPExceptionMapper implements ExceptionMapper<Exception> {
     }
     if (ex instanceof OTPRequestTimeoutException) {
       return Response
-        .status(
-          Response.Status.fromStatusCode(OtpHttpStatus.STATUS_UNPROCESSABLE_ENTITY.statusCode())
-        )
+        .status(OtpHttpStatus.STATUS_UNPROCESSABLE_ENTITY.statusCode())
         .entity("OTP API Processing Timeout")
         .type("text/plain")
         .build();
@@ -60,7 +58,7 @@ public class OTPExceptionMapper implements ExceptionMapper<Exception> {
     LOG.error("Unhandled exception", ex);
     // Return the short form message to the client
     return Response
-      .status(Response.Status.INTERNAL_SERVER_ERROR)
+      .serverError()
       .entity(ex.toString() + " " + ex.getMessage())
       .type("text/plain")
       .build();