+++ src/main/java/org/opentripplanner/standalone/config/routerconfig/updaters/WFSNotePollingGraphUpdaterConfig.java
@@ -2,6 +2,7 @@ package org.opentripplanner.standalone.config.routerconfig.updaters;
 
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.NA;
 
+import java.time.Duration;
 import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
 import org.opentripplanner.updater.street_note.WFSNotePollingGraphUpdaterParameters;
 
@@ -12,7 +13,7 @@ public class WFSNotePollingGraphUpdaterConfig {
       configRef,
       c.of("url").since(NA).summary("TODO").asString(),
       c.of("featureType").since(NA).summary("TODO").asString(),
-      c.of("frequencySec").since(NA).summary("TODO").asInt(60)
+      c.of("frequency").since(NA).summary("TODO").asDuration(Duration.ofMinutes(1))
     );
   }
 }