+++ src/main/java/org/opentripplanner/updater/vehicle_position/PollingVehiclePositionUpdater.java
@@ -53,8 +53,8 @@ public class PollingVehiclePositionUpdater extends PollingGraphUpdater {
       );
 
     LOG.info(
-      "Creating vehicle position updater running every {} seconds : {}",
-      pollingPeriodSeconds(),
+      "Creating vehicle position updater running every {}: {}",
+      pollingPeriod(),
       vehiclePositionSource
     );
   }